warnings

+++ Kernel/exceptions.c
@@ -1,4 +1,5 @@
 #include "include/exceptionsEngine.h"
+
 static const char *registerNames[REGS_CANT + 1] = {
             "RAX", "RBX", "RCX", "RDX", "RSI", "RDI", "RBP", "RSP", "R8 ", "R9 ", "R10", "R11", "R12", "R13",
             "R14", "R15", "RIP"
@@ -18,7 +19,7 @@ static void reportRegs(long regs[], char fd){
 }
 
 void exceptionDispatcher(int exception, long regs[],char * contextOwner) {
-	int fd = getFD(*contextOwner);
+	int fd = getFD((int)*contextOwner);
 	fdClear(fd);
 
 	ncClearHeader();
